Chatbot/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4==4.12.3",
    "fastapi==0.115.6",
    "groq==0.13.1",
    "httpx==0.28.1",
    "langchain==0.3.14",
    "langchain-community==0.3.14",
    "langchain-groq==0.2.2",
    "langgraph==0.2.62",
    "lxml==5.3.0",
    "numpy>=2.4.2",
    "pydantic==2.10.4",
    "python-dotenv==1.0.1",
    "qdrant-client==1.12.1",
    "selenium==4.27.1",
    "sentence-transformers==3.3.1",
    "transformers==4.47.1",
    "uvicorn[standard]==0.34.0",
    "webdriver-manager==4.0.2",
]
